feat: move backup util to pkg/util/backup
